1-N

+++ app/Http/Controllers/MovieController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\MovieRequest;
 use App\Models\Movie;
 use Illuminate\Http\Request;
+use App\Http\Requests\MovieRequest;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class MovieController extends Controller
@@ -12,7 +13,7 @@ class MovieController extends Controller
     public function __construct()
     {
         // Protegge tutte le rotte tranne la lista pubblica dei film
-        $this->middleware('auth')->except('movieList', 'movieDetail');
+        $this->middleware('auth')->except('movieList');
     }
 
     // Lista tutti i film
@@ -43,6 +44,7 @@ class MovieController extends Controller
             'year' => $request->year,
             'plot' => $request->plot,
             'img' => $request->file('img')->store('images', 'public'),
+            'user_id' => Auth::user()->id
         ]);
 
         return redirect()->route('movies.index')->with('successMessage', "Il tuo film è stato caricato correttamente");
@@ -51,40 +53,52 @@ class MovieController extends Controller
     // Mostra il form per modificare un film esistente
     public function edit(Movie $movie)
     {
-        return view('movie.edit', compact('movie'));
+        if ($movie->user_id == Auth::user()->id) {
+            return view('movie.edit', compact('movie'));
+        } else {
+            return redirect()->route('homepage')->with('errorMessage', 'Non hai i permessi di vedere questa pagina');
+        }
     }
 
     // Aggiorna il film nel database
     public function update(Request $request, Movie $movie)
     {
-        $movie->update([
-            'title' => $request->title,
-            'director' => $request->director,
-            'year' => $request->year,
-            'plot' => $request->plot,
-        ]);
+        if ($movie->user == Auth::user()->id) {
+            $movie->update([
+                'title' => $request->title,
+                'director' => $request->director,
+                'year' => $request->year,
+                'plot' => $request->plot,
+            ]);
 
-        // Se viene caricata una nuova immagine, aggiorna anche quella
-        if ($request->hasFile('img')) {
-            if ($movie->img) {
-                Storage::disk('public')->delete($movie->img);
+            // Se viene caricata una nuova immagine, aggiorna anche quella
+            if ($request->hasFile('img')) {
+                if ($movie->img) {
+                    Storage::disk('public')->delete($movie->img);
+                }
+                $movie->img = $request->file('img')->store('images', 'public');
+                $movie->save();
             }
-            $movie->img = $request->file('img')->store('images', 'public');
-            $movie->save();
-        }
 
-        return redirect()->route('movies.show', $movie)->with('successMessage', 'Film aggiornato con successo');
+            return redirect()->route('movies.show', $movie)->with('successMessage', 'Film aggiornato con successo');
+        } else {
+            return redirect()->route('homepage')->with('errorMessage', 'Non hai i permessi di vedere questa pagina');
+        }
     }
 
     // Elimina un film dal database e cancella l'immagine associata
     public function destroy(Movie $movie)
     {
-        if ($movie->img) {
-            Storage::disk('public')->delete($movie->img);
+        if ($movie->user_id == Auth::user()->id) {
+            if ($movie->img) {
+                Storage::disk('public')->delete($movie->img);
+            }
+
+            $movie->delete();
+
+            return redirect()->route('movies.index')->with('successMessage', 'Film eliminato correttamente');
+        } else {
+            return redirect()->route('homepage')->with('errorMessage', 'Non hai i permessi di vedere questa pagina');
         }
-
-        $movie->delete();
-
-        return redirect()->route('movies.index')->with('successMessage', 'Film eliminato correttamente');
     }
 }
